refactor: move job and sequence flow logic in behavior classes
Move the Job- and Sequence Flow-specific migration
logic from the migration processor to dedicated
behavior classes.

The goal is to improve the readability and
maintainability of the migration processor.
